+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a warning in case too many or too few questions were loaded.
+
 
 ## [0.1.4]  - 2024-01-17